Move sphinxthread to utils.
